backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
asyncpg
python-dotenv
lxml
pandas
requests
tqdm